pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "chalcogenide-perovskite-screening"
version = "1.0.0"
description = "ML-guided screening of chalcogenide perovskites as solar energy materials"
authors = [
  { name = "Diego A. Garzón", email = "diego.garzon@inl.int" },
  { name = "Lauri Himanen" },
  { name = "Luisa Andrade" },
  { name = "Sascha Sadewasser" },
  { name = "José A. Márquez" },
]
license = { file = "LICENSE" }
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Chemistry",
]
requires-python = ">=3.8,<=3.8.20"

dependencies = [
    # Core scientific
    "numpy>=1.24,<2",
    "pandas>=2.0",
    "scipy>=1.10",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.13",
    "statsmodels>=0.14",
    # Materials science
    "pymatgen>=2023.8",
    "ase>=3.23",
    "spglib>=2.4",
    "mp-api>=0.35",
    # ML / deep learning
    "torch>=2.4",
    "crabnet>=2.0.8",
    # Visualization
    "plotly>=5.6",
    "kaleido>=0.2",
    "pymatviz>=0.6",
    # Utilities
    "loguru",
    "python-dotenv",
    "tqdm",
    "typer>=0.9",
    "openpyxl",
    "uncertainties",
    "pydantic>=1.10,<2",
    "tabulate",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-timeout",
    "ruff",
    "black",
]
notebooks = [
    "ipykernel>=6.29",
    "ipywidgets",
]

# NOTE: sissopp is a compiled Fortran package NOT available on PyPI.
# Install manually from https://github.com/rouyang2017/SISSO if you need
# to re-derive SISSO features. Cached features ship in data/interim/features_sisso.csv.

[tool.black]
line-length = 99
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
)/
'''

[tool.ruff]
line-length = 99
exclude = [".git", "notebooks", "references", "models", "data"]

[tool.ruff.lint]
ignore = ["E731", "E501", "C901"]

[tool.ruff.lint.isort]
known-first-party = ["chalcogenide_perovskite_screening"]
force-sort-within-sections = true
